+++ src/Evaluator/EvalError.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml;
+namespace Serhii\GoodbyeHtml\Evaluator;
 
 use Serhii\GoodbyeHtml\Ast\Node;
 use Serhii\GoodbyeHtml\Ast\VariableExpression;
@@ -10,12 +10,12 @@ use Serhii\GoodbyeHtml\Obj\ErrorObj;
 use Serhii\GoodbyeHtml\Obj\Obj;
 use Serhii\GoodbyeHtml\Obj\ObjType;
 
-readonly class ErrorMessage
+readonly class EvalError
 {
     public static function wrongArgumentType(string $type, ObjType $expect, Obj $actual): ErrorObj
     {
         return new ErrorObj(sprintf(
-            '[ERROR] "%s" is not allowed argument type for "%s", expected "%s"',
+            '[EVAL_ERROR] "%s" is not allowed argument type for "%s", expected "%s"',
             $actual->type()->value,
             $type,
             $expect->value,
@@ -24,16 +24,16 @@ readonly class ErrorMessage
 
     public static function unknownType(Node $node): ErrorObj
     {
-        return new ErrorObj('[ERROR] unknown type: "' . get_class($node) . '"');
+        return new ErrorObj('[EVAL_ERROR] unknown type: "' . get_class($node) . '"');
     }
 
     public static function unknownOperator(string $operator, Obj $right): ErrorObj
     {
-        return new ErrorObj(sprintf('[ERROR] unknown operator "%s%s"', $operator, $right->type()->value));
+        return new ErrorObj(sprintf('[EVAL_ERROR] unknown operator "%s%s"', $operator, $right->type()->value));
     }
 
     public static function variableIsUndefined(VariableExpression $node): ErrorObj
     {
-        return new ErrorObj(sprintf('[ERROR] variable "$%s" is undefined', $node->value));
+        return new ErrorObj(sprintf('[EVAL_ERROR] variable "$%s" is undefined', $node->value));
     }
 }
